Move injectable lazy document access to top level and rename to VespaDocumentAccess
